Update README.md
Premier League Website Redesign is a project that I undertook to create a more user-friendly version of the Premier League website. While the official site is rich in content, I found it a bit too complex for quick navigation and straightforward information retrieval. This project simplifies the browsing experience, focusing on delivering key information in a more organized and accessible manner. The redesign includes sections for league standings, match schedules, team information, player statistics, and more, all presented in a clean and easy-to-navigate format. The site was built using HTML, CSS, and JavaScript, ensuring it remains fast and responsive while offering all essential details about the Premier League.